feat: 제품 기능 구현 완료 (#77) chore: 타임리프 설정 비활성화 (#77) feat: 예외처리에 추가 예외 추가 (#77) refactor: findProduct 중복으로 인해 findProductId로 변경 (#77)

+++ server/src/main/java/seb/project/Codetech/product/service/ProductService.java
@@ -7,27 +7,35 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import seb.project.Codetech.global.exception.BusinessLogicException;
+import seb.project.Codetech.global.exception.ExceptionCode;
 import seb.project.Codetech.global.page.PageInfo;
 import seb.project.Codetech.product.entity.Product;
 import seb.project.Codetech.product.repository.ProductRepository;
+import seb.project.Codetech.user.entity.User;
+import seb.project.Codetech.user.repository.UserRepository;
 
 @Service
 public class ProductService {
 
 	private final ProductRepository productRepository;
+	private final UserRepository userRepository;
 
-	public ProductService(ProductRepository productRepository) {
+	public ProductService(ProductRepository productRepository, UserRepository userRepository) {
 		this.productRepository = productRepository;
+		this.userRepository = userRepository;
 	}
 
-	public Product createProduct(Product product) {
+	public Product createProduct(String email, Product product) {
+		product.setWriter(findUser(email).getNickname()); // 작성자를 기록한다 한번 기록하면 수정이 불가능해진다.
+		product.setModifier(findUser(email).getNickname()); // 수정자도 동시에 작성자로 기록한다.
 		return productRepository.save(product);
 	}
 
-	public Product modifyProduct(Product product) {
-		Product findProduct = productRepository.findById(product.getId())
-			.orElseThrow(() -> new RuntimeException("제품 정보를 찾을 수 없습니다."));
+	public Product modifyProduct(String email, Product product) {
+		Product findProduct = findProductId(product.getId());
 		Optional.ofNullable(product.getName()).ifPresent(findProduct::setName);
+		Optional.ofNullable(findUser(email).getNickname()).ifPresent(findProduct::setModifier); // 수정한 유저 이름을 기록한다.
 		Optional.ofNullable(product.getDetail()).ifPresent(findProduct::setDetail);
 		Optional.ofNullable(product.getImage()).ifPresent(findProduct::setImage);
 		return productRepository.save(findProduct);
@@ -35,7 +43,7 @@ public class ProductService {
 
 	public Product findProduct(Product product) {
 		return productRepository.findById(product.getId())
-			.orElseThrow(() -> new RuntimeException("제품 정보를 찾을 수 없습니다."));
+			.orElseThrow(() -> new BusinessLogicException(ExceptionCode.PRODUCT_NOT_FOUND));
 	}
 
 	public Page<Product> findAllProduct(PageInfo.Request request) {
@@ -43,12 +51,21 @@ public class ProductService {
 			PageRequest.of(request.getPage(), request.getSize(), Sort.by("id").descending()));
 	}
 
-	public void removeProduct(Product product) {
+	public void removeProduct(String email, Product product) {
+
+		if (!findUser(email).getRoles().contains("ADMIN")) { // 작성된 제품 정보는 관리자만 삭제할 수 있다.
+			throw new BusinessLogicException(ExceptionCode.PRODUCT_NOT_DELETE);
+		}
 		productRepository.delete(product);
 	}
 
-	public Product findProduct(Long id) {
+	public Product findProductId(Long id) { // 제품 정보를 가져오기 위한 메소드
 		return productRepository.findById(id)
-			.orElseThrow(() -> new RuntimeException("제품 정보를 찾을 수 없습니다."));
+			.orElseThrow(() -> new BusinessLogicException(ExceptionCode.PRODUCT_NOT_FOUND));
+	}
+
+	public User findUser(String email) { // 유저 데이터를 가져오기 위한 메소드
+		return userRepository.findByEmail(email)
+			.orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND));
 	}
 }
